Live strategy

+++ user_data/strategies/super_multi_factors_strategy.py
@@ -29,7 +29,7 @@ class SuperMultiFactorsStrategy(IStrategy):
     factors_validated_time = None
     process_only_new_candles = True
     can_short = True
-    startup_candle_count = 800
+    startup_candle_count = 3000
     use_custom_stoploss = True
     is_validating_factors = False
 
@@ -63,37 +63,44 @@ class SuperMultiFactorsStrategy(IStrategy):
 
     def bot_loop_start(self, current_time: datetime, **kwargs) -> None:
         # Check whether factor needs to re-validate
-        if self.factors_validated_time is not None and current_time - datetime.timedelta(hours=24) < self.factors_validated_time:
+        if self.factors_validated_time is not None:
             return
 
-        # Check whether data collected ready
-        if not self.is_pair_data_prepared():
-            return
-        # Do validation again
-        logger.info("Starting re-validating factors")
-        self.is_validating_factors = True
-        self.factor_generator_5m.validate_factors(self.pair_data_dict_5m)
-        logger.info(f"5m timeframe factors {self.factor_generator_5m.factor_methods}")
-        self.factor_generator_15m.validate_factors(self.pair_data_dict_15m)
-        logger.info(f"15m timeframe factors {self.factor_generator_15m.factor_methods}")
-        self.factor_generator_1h.validate_factors(self.pair_data_dict_1h)
-        logger.info(f"1h timeframe factors {self.factor_generator_1h.factor_methods}")
+        # We don't revalidate in the live
+        if self.dp.runmode != 'live':
+            # Check whether data collected ready
+            if not self.is_pair_data_prepared():
+                return
+            # Do validation again
+            logger.info("Starting re-validating factors")
+            self.is_validating_factors = True
+            self.factor_generator_5m.validate_factors(self.pair_data_dict_5m)
+            logger.info(f"5m timeframe factors {self.factor_generator_5m.factor_methods}")
+            self.factor_generator_15m.validate_factors(self.pair_data_dict_15m)
+            logger.info(f"15m timeframe factors {self.factor_generator_15m.factor_methods}")
+            self.factor_generator_1h.validate_factors(self.pair_data_dict_1h)
+            logger.info(f"1h timeframe factors {self.factor_generator_1h.factor_methods}")
 
-        # Save the validated time and set factor_validation_ready to True
-        self.factors_validated_time = current_time
-        self.is_validating_factors = False
+            # Save the validated time and set factor_validation_ready to True
+            self.factors_validated_time = current_time
+            self.is_validating_factors = False
+            # Set freqtrade ai model expire
+            for pair in self.dp.current_whitelist():
+                if self.freqai.dd.pair_dict.get(pair):
+                    self.freqai.dd.pair_dict[pair] = self.freqai.dd.empty_pair_dict.copy()
 
-        # Set freqtrade ai model expire
-        for pair in self.dp.current_whitelist():
-            if self.freqai.dd.pair_dict.get(pair):
-                self.freqai.dd.pair_dict[pair] = self.freqai.dd.empty_pair_dict.copy()
+            # Remove model file from disk
+            model_folders = [x for x in self.freqai.dd.full_path.iterdir() if x.is_dir()]
+            for model_folder in model_folders:
+                shutil.rmtree(model_folder)
 
-        # Remove model file from disk
-        model_folders = [x for x in self.freqai.dd.full_path.iterdir() if x.is_dir()]
-        for model_folder in model_folders:
-            shutil.rmtree(model_folder)
-
-        logger.info("Re-validating factors succeed")
+            logger.info("Re-validating factors succeed")
+        else:
+            logger.info("Running strategy in live, loading factors from csv")
+            self.factor_generator_5m.load_factors_from_file()
+            self.factor_generator_15m.load_factors_from_file()
+            self.factor_generator_1h.load_factors_from_file()
+            self.factors_validated_time = current_time
 
     def is_pair_data_prepared(self):
         all_data_collected = all((pair in self.pair_data_dict_15m) and (pair in self.pair_data_dict_5m) and (pair in self.pair_data_dict_1h)
@@ -116,9 +123,10 @@ class SuperMultiFactorsStrategy(IStrategy):
     def feature_engineering_expand_basic(
         self, dataframe: DataFrame, metadata: dict, **kwargs
     ) -> DataFrame:
-        dataframe["%-pct-change"] = dataframe["close"].pct_change()
-        dataframe["%-raw_volume"] = dataframe["volume"]
-        dataframe["%-raw_price"] = dataframe["close"]
+        if self.factors_validated_time is not None and not self.is_validating_factors:
+            dataframe["%-pct-change"] = dataframe["close"].pct_change()
+            dataframe["%-raw_volume"] = dataframe["volume"]
+            dataframe["%-raw_price"] = dataframe["close"]
         return dataframe
 
     def feature_engineering_standard(
@@ -137,12 +145,14 @@ class SuperMultiFactorsStrategy(IStrategy):
 
     @informative('15m')
     def populate_indicators_15m(self, dataframe: pd.DataFrame, metadata: dict) -> pd.DataFrame:
-        self.update_pair_data(metadata['pair'], '15m', dataframe)
+        if self.dp.runmode != 'live':
+            self.update_pair_data(metadata['pair'], '15m', dataframe)
         return dataframe
 
     @informative('1h')
     def populate_indicators_1h(self, dataframe: pd.DataFrame, metadata: dict) -> pd.DataFrame:
-        self.update_pair_data(metadata['pair'], '1h', dataframe)
+        if self.dp.runmode != 'live':
+            self.update_pair_data(metadata['pair'], '1h', dataframe)
         return dataframe
 
     def populate_indicators(self, dataframe: DataFrame, metadata: dict) -> DataFrame:  # noqa: C901
@@ -171,7 +181,8 @@ class SuperMultiFactorsStrategy(IStrategy):
         # TEMA - Triple Exponential Moving Average
         dataframe["tema"] = ta.TEMA(dataframe, timeperiod=9)
 
-        self.update_pair_data(metadata['pair'], '5m', dataframe)
+        if self.dp.runmode != 'live':
+            self.update_pair_data(metadata['pair'], '5m', dataframe)
         return dataframe
 
     def populate_entry_trend(self, df: DataFrame, metadata: dict) -> DataFrame:
@@ -182,16 +193,13 @@ class SuperMultiFactorsStrategy(IStrategy):
             (
                 # Signal: RSI crosses above 30
                 (qtpylib.crossed_above(df["rsi"], self.buy_rsi.value))
-                &
-                (df["tema"] < df["bb_middleband"])  # Guard: tema below BB middle
+                & (df["tema"] <= df["bb_middleband"])  # Guard: tema below BB middle
                 & (df["tema"] > df["tema"].shift(1))  # Guard: tema is raising
                 & (df["volume"] > 0)  # Make sure Volume is not 0
                 & (df["do_predict"] == 1)  # Make sure Freqai is confident in the prediction
                 &
                 # Only enter trade if Freqai thinks the trend is in this direction
                 (df["&s-up_or_down"] == "up")
-                # &
-                # (self.factors_validated_time is not None)
             ),
             "enter_long",
         ] = 1
@@ -199,17 +207,14 @@ class SuperMultiFactorsStrategy(IStrategy):
         df.loc[
             (
                 # Signal: RSI crosses above 70
-                (qtpylib.crossed_below(df["rsi"], self.short_rsi.value))
-                &
-                (df["tema"] > df["bb_middleband"])  # Guard: tema above BB middle
+                (qtpylib.crossed_above(df["rsi"], self.short_rsi.value))
+                & (df["tema"] > df["bb_middleband"])  # Guard: tema above BB middle
                 & (df["tema"] < df["tema"].shift(1))  # Guard: tema is falling
                 & (df["volume"] > 0)  # Make sure Volume is not 0
                 & (df["do_predict"] == 1)  # Make sure Freqai is confident in the prediction
                 &
                 # Only enter trade if Freqai thinks the trend is in this direction
                 (df["&s-up_or_down"] == "down")
-                # &
-                # (self.factors_validated_time is not None)
             ),
             "enter_short",
         ] = 1
@@ -217,14 +222,16 @@ class SuperMultiFactorsStrategy(IStrategy):
         return df
 
     def populate_exit_trend(self, df: DataFrame, metadata: dict) -> DataFrame:
+        if self.factors_validated_time is None or self.is_validating_factors:
+            return df
+
         df.loc[
             (
                 # Signal: RSI crosses above 70
-                    (qtpylib.crossed_above(df["rsi"], self.sell_rsi.value))
-                    &
-                    (df["tema"] > df["bb_middleband"])  # Guard: tema above BB middle
-                    & (df["tema"] < df["tema"].shift(1))  # Guard: tema is falling
-                    & (df["volume"] > 0)  # Make sure Volume is not 0
+                (qtpylib.crossed_above(df["rsi"], self.sell_rsi.value))
+                & (df["tema"] > df["bb_middleband"])  # Guard: tema above BB middle
+                & (df["tema"] < df["tema"].shift(1))  # Guard: tema is falling
+                & (df["volume"] > 0)  # Make sure Volume is not 0
             ),
             "exit_long",
         ] = 1
@@ -232,45 +239,45 @@ class SuperMultiFactorsStrategy(IStrategy):
         df.loc[
             (
                 # Signal: RSI crosses above 30
-                    (qtpylib.crossed_below(df["rsi"], self.exit_short_rsi.value))
-                    &
-                    # Guard: tema below BB middle
-                    (df["tema"] < df["bb_middleband"])
-                    & (df["tema"] > df["tema"].shift(1))  # Guard: tema is raising
-                    & (df["volume"] > 0)  # Make sure Volume is not 0
+                (qtpylib.crossed_above(df["rsi"], self.exit_short_rsi.value))
+                &
+                # Guard: tema below BB middle
+                (df["tema"] <= df["bb_middleband"])
+                & (df["tema"] > df["tema"].shift(1))  # Guard: tema is raising
+                & (df["volume"] > 0)  # Make sure Volume is not 0
             ),
             "exit_short",
         ] = 1
 
         return df
 
-    def confirm_trade_entry(
-        self,
-        pair: str,
-        order_type: str,
-        amount: float,
-        rate: float,
-        time_in_force: str,
-        current_time: datetime,
-        entry_tag: Optional[str],
-        side: str,
-        **kwargs,
-    ) -> bool:
-        is_short = True if side == "short" else False
-        # Max trade count in each side
-        can_open_trades_len = int(self.config["max_open_trades"]) // 2
-        num_shorts, num_longs, has_pair_pos = self.get_num_pos(pair, is_short)
-        # Already has position of this pair
-        if has_pair_pos:
-            return False
-
-        if side == "long" and num_longs >= can_open_trades_len:
-            return False
-
-        if side == "short" and num_shorts >= can_open_trades_len:
-            return False
-
-        return True
+    # def confirm_trade_entry(
+    #     self,
+    #     pair: str,
+    #     order_type: str,
+    #     amount: float,
+    #     rate: float,
+    #     time_in_force: str,
+    #     current_time: datetime,
+    #     entry_tag: Optional[str],
+    #     side: str,
+    #     **kwargs,
+    # ) -> bool:
+    #     is_short = True if side == "short" else False
+    #     # Max trade count in each side
+    #     can_open_trades_len = int(self.config["max_open_trades"]) // 2
+    #     num_shorts, num_longs, has_pair_pos = self.get_num_pos(pair, is_short)
+    #     # Already has position of this pair
+    #     if has_pair_pos:
+    #         return False
+    #
+    #     if side == "long" and num_longs >= can_open_trades_len:
+    #         return False
+    #
+    #     if side == "short" and num_shorts >= can_open_trades_len:
+    #         return False
+    #
+    #     return True
 
     def custom_stoploss(
         self,
@@ -284,13 +291,19 @@ class SuperMultiFactorsStrategy(IStrategy):
     ) -> Optional[float]:
         dataframe, _ = self.dp.get_analyzed_dataframe(pair, self.timeframe)
         last_candle = dataframe.iloc[-1].squeeze()
-        atr = last_candle["atr"]
+        pos_minutes = (current_time - trade.open_date_utc.replace(second=0, microsecond=0)).total_seconds() // 60
 
-        if current_profit > 3 * atr:
+        atr_ratio = 3
+        # if pos_minutes >= 120:
+        #     if (trade.is_short and last_candle['&s-up_or_down'] == 'up') or (not trade.is_short and last_candle['&s-up_or_down'] == 'down'):
+        #         atr_ratio = 2
+
+        atr = last_candle["atr"]
+        if current_profit > atr_ratio * atr:
             desired_stoploss = current_profit / 2
             return max(min(desired_stoploss, 0.01), 0.005)
 
-        if current_profit < -3 * atr:
+        if current_profit < -atr_ratio * atr:
             return -atr
